tutorial3: info&answers
CSC 209 tutorial 03 exercises, week of 25 January 2016

Please begin by running "/student/csc209/present". This records your attendance and is required for you to get the marks for the lab.

Today's tutorial is about writing "man pages".

First, consider the device file /dev/urandom. You can read about it via "man urandom", but basically, it is a source of random numbers. You won't want to cat it, because it's an endless stream, and because it's all binary -- each byte is a random value between 0 and 255 inclusive.

It's easy to work with from a C program, but not so easy to do so from shell commands. So I wrote a little program named "firstbyte" which outputs the first byte of a file, and as a decimal number. You can find it in /student/csc209/tut/03/firstbyte.c 
It takes an optional command-line argument which is a file name to read instead of standard input. Without any command-line arguments it reads its standard input.

But you aren't expected to be able to read that C code yet. Instead, you can simply run the compiled version, which you can find in /student/csc209/tut/03/firstbyte

First, play with the program a little and make sure you know what it does and how to use it. Here are a few sample commands to try (try the first two a few times in a row (using the up-arrow key to get a previous command back) because they will produce different numbers each time!):

	/student/csc209/tut/03/firstbyte /dev/urandom
	/student/csc209/tut/03/firstbyte </dev/urandom
	echo hello | /student/csc209/tut/03/firstbyte
Continue to experiment for a minute or two. It's straightforward, but make sure you know it thoroughly, with filename arguments, and with stdin. Note that interactive use of stdin looks rather strange; so instead try it more along the lines of the above commands, with a file redirected as stdin, or with its input coming from a pipe.
Second, please skim the detailed document "How to write a manual page" which you can find in https://cs.utm.utoronto.ca/~ajr/209/manpage.pdf as a PDF file. Focus primarily on the beginning of the document; get an idea of the document and become able to find things in it which you will need in writing your man page.

Third, read my additional notes on man-page writing, all of which I think are relevant to the current exercise, at https://cs.utm.utoronto.ca/~ajr/209/manpage-ajr.html 
Regarding the second "bullet point", in particular note that you will not be documenting things like how if you use firstbyte interactively in the shell and type more than one byte, the rest of the input is a shell command -- this is just about how stdin works and is not a firstbyte topic. Just say that it only reads one byte.

Finally, write a man page for the "firstbyte" program. Call your file "firstbyte.1" -- that indicates that this is a man page for volume one of "The Unix Programmer's Manual".

This man page will be graded in detail; unlike some labs this term you can't expect to get 100% just for handing in something reasonable. Pay attention to detail, brevity, and clarity.

Useful command to format the man page and display it on the screen just like how the 'man' command does for installed man pages:

	nroff -man firstbyte.1 | less