tutorial5: info&answers
CSC 209 tutorial 05 exercises, week of 8 February 2016

Please begin by running "/student/csc209/present". This records your attendance and is required for you to get the marks for the lab.
1. Demonstrations of selected system calls

Please look at the sample programs in /student/csc209/tut/05.
They are all compiled there in that directory for you, but you can also copy the .c files to your own directory to modify them to test variations.

First, examine the program /student/csc209/tut/05/stat.c 
The stat() and/or lstat() system calls will be needed for assignment two.

stat.c illustrates calling stat() on an argument file name and extracting various information about the file. The data returned from stat() is basically an "ls -l".

This data is returned in a "struct" because there are too many values for it to be feasible to do it any other way. A "struct" is basically a dataless class -- it is an aggregate data type containing arbitrary other data types. Some of the contents of this struct are output by the example stat.c. The semantics of the "." in C are not the same as in Java or Python, but to some extent you can ignore the distinction for now, and we'll talk about structs fully in a subsequent lecture.

Note that if you run stat (or my stat.c) on a symlink, you get information about the pointed-to file (try it! -- e.g. create a file named "foo" and then create a symlink to it with "ln −s foo bar"). This is the difference between stat() and lstat(). Examine and run lstat.c. (And try "diff stat.c lstat.c"!) 
(Note: If you just type "stat" as the command name, intending to run a file "stat" out of the current directory, you may instead get /usr/bin/stat, depending on how your PATH is set up. I suggest getting into the habit of typing "./file" to run a program in the current directory, and omitting dot from your PATH.)

Normally we want to "follow" symlinks in the way that stat() does. But for some purposes we need to see the actual symlink. For example, to write "ls -l". For another example, your assignment two directory-traversing code (part 3 of assignment two) will need to use lstat() rather than stat() to traverse a directory tree properly, so that symlinks don't send it off in strange directions (possibly even infinite loops).

Three of the data values in the "struct stat" are time values. A time value in unix/linux is just a single integer which is the number of seconds since 00:00 1 Jan 1970 UTC, so it's particularly easy to work with, although difficult to convert into a string suitable for output to the user, but we have library routines for the latter. For some examples of manipulating time values you can see /student/csc209/tut/05/time.c , although you don't need that for assignment two nor this tutorial (all you need to do in this tutorial is to compare the time values for order).

In case you're not completely clear on strcmp() (which is used in assignment two, although it's not a system call), you can look at /student/csc209/tut/05/strcmp.c, which is just a command-line front end to strcmp(). Run this program with various strings in argv[1] and argv[2].

Finally, for directory traversal you will need to read a directory. The program /student/csc209/tut/05/readdir.c illustrates using opendir(), readdir(), and closedir(). The struct returned by readdir() contains all of the information in one entry in the directory -- which is merely an inode number and a name.

In stat() and lstat(), we passed in a pointer to a struct which the library routine filled up; whereas in this case, we get a pointer back from readdir(). Thus unlike in stat.c, here we're usually dealing with a variable which is a pointer to a struct, rather than the struct itself.

So we could write things like "(*p).d_name", first dereferencing the pointer-to-struct, then accessing a member of the struct. This notation is cumbersome, though, and the parentheses are necessary because of the operator precedence rules. So there's a special notation in C which looks like a right-pointing arrow and is formed from the two characters '-' and '>". "a->b" is simply defined as "(*a).b". Thus instead of the cumbersome "(*p).d_name", we can write "p->d_name". (It makes more of a difference in typing than in reading it! It also makes more of a difference when there are multiple levels of pointers to structs. And we'll see all of this more in future lectures.)

2. To submit

Write a C program which insists on having two arguments (i.e. argc==3). Both arguments are expected to be file path names. Call stat() on each of them and output the file path name which is newer (i.e. for which the st_mtime struct member is numerically greater). Note that there is no other output other than this file path name (and a newline character).
You want to take care in copying any portions of the supplied stat.c. In general, it's asking for trouble to copy code you don't want or understand; write your own program and make sure you know what's in there.

Call your program tut5.c and submit it with

        /student/csc209/submit tut5 tut5.c
, by the end of the day; and no matter when you submit it, don't forget to run /student/csc209/present during the tutorial time, on the console of a tutorial lab workstation.
(Other 'submit' commands are described at the end of the notes for tutorial one.)

Your program must compile on the UTM linux machines with "gcc -Wall" with no errors or warning messages.

Q: Do we need to check argc and output a usage message?

A: Yes! Always! And the usage message must be in the standard format (see the SYNOPSIS line from the man page exercise in tutorial three), and to stderr, and you must exit with exit status 1 in case of error or 0 in the normal case.

Note: Since we hadn't talked about output to other files yet, outputting the usage message to stdout was accepted for tutorial four... but that won't be accepted for this or future tutorials, nor for assignments. Output usage messages to stderr as we always do for error messages.