tutorial6: info&answers
CSC 209 tutorial 06 exercises, week of 22 February 2016

Please begin by running "/student/csc209/present". This records your attendance and is required for you to get the marks for the lab.
1. Files

Take the rot.c program from assignment one (either your solution or my solution in /student/csc209/a1/rot.c) and make it function like a normal "zero or more file names" unix tool.
That is, argc is required to be at least 2, and argv[1] is the rotation amount; but after that, if argc is 2 it processes the standard input (as before), otherwise it loops through the remaining arguments, fopen()s each one, process it, fclose().

So instead of getchar() in your loop, you use getc(fp). You still use putchar().

Avoid duplicating code. When you go to process stdin, you could do so with the same code (probably in a separate function) by making fp be the value stdin (defined in stdio.h).

2. Strings

2.1. Add the loop described below to the following code (available in /student/csc209/tut/06/starter.c, or copy from this web page):
int main()
{
    char s[27];
    LOOP GOES HERE
    printf("%s\n", s);
    return(0);
}
The program should output "ABCDEFGHIJKLMNOPQRSTUVWXYZ" (without the quotes), but don't use any of the str* functions (e.g. strcpy). Create the string 's' by setting the items in the array to the successive letters of the alphabet using a simple for loop by doing arithmetic on character values, and then setting s[26] to zero.

2.2. Given an int value between 10 and 99 inclusive (e.g. start your code with "int x = 38;" for a test), write code to convert this to a three-character array which is a string. The zeroth character of your string is the number div 10 + '0'; character one of your string is the number mod 10 + '0'; and the array element two is the terminating zero. Note the difference between 0 and '0' (the latter is 48 in ASCII).

3. To submit: string processing

A "palindrome" is a coherent phrase which consists of the same letters and digits forwards and backwards, although the spacing and capitalization may be different. Some examples:
	Madam, I'm Adam.
	A man, a plan, a canal, Panama.
	Doc, note: I dissent!  A fast never prevents a fatness.  I diet on cod.
	Won 12 tons?  Not 21 now.
Write a C program named "ispalindrome.c" which insists on having one argument (i.e. argc == 2). It checks its argument string to see whether it's a palindrome, and exits with exit status 0 if it is a palindrome, or 1 if it is not. Similar to the test program, there is no output, unless there is a usage error.

An example of using ispalindrome is in the "findpal" program in assignment one, a solution to which (in which you'd have to modify the ispalindrome invocation to run your own, if you want to test them together) can be found in /student/csc209/a1/findpal 
(and regarding that solution, also see /student/csc209/a1/findpal-README , although that's not relevant to this tutorial)

Implementation technique: Start a pointer-to-char at the beginning of the string and another pointer-to-char at the end of the string. (Find the end of the string with a simple linear search algorithm to find the zero byte.) The pointer which started at the beginning of the string gets incremented and the pointer which started at the end of the string gets decremented, until they meet or pass. 
So your loop will look something like "while (p < q)".

Note: Since this exercise is about strings and pointers, you must use this pointer strategy to get the point for the lab. That is to say, your access to a character in the string has to look more like "*p" than like "s[i]". (And "*(s+i)" doesn't count as the pointer strategy, that's the same as s[i] (by definition). You have to use a pointer variable which you are stepping through the string without indexing.)

Note relevant functions available after "#include <ctype.h>": isalnum(c) returns whether the character c is alphanumeric (you should skip all non-alphanumeric characters in making your comparison), and tolower(c) returns a lower-case version of the character c if it's an upper-case letter (e.g. tolower('E') returns 'e'), and returns the character itself otherwise (e.g. tolower('e') returns 'e', and tolower('2') returns '2').

As usual, submit your program with

        /student/csc209/submit tut6 ispalindrome.c
, by the end of the day, and no matter when you submit it, don't forget to run /student/csc209/present during the tutorial time, on the console of a tutorial lab workstation.
(Other 'submit' commands are described at the end of the notes for tutorial one.)

Your program must compile on the UTM linux machines with "gcc -Wall" with no errors or warning messages.

Q: Do we need to check argc and output a usage message?

A: Yes! Always! And the usage message must be in the standard format (see the SYNOPSIS line from the man page exercise in tutorial three), and to stderr.