tutorial8: info&answers
CSC 209 tutorial 08 exercises, week of 7 March 2016

Please begin by running "/student/csc209/present". This records your attendance and is required for you to get the marks for the lab.
1. fork exercise

Write a program which creates five child processes, numbered 0 through 4. Each child process computes the sum of the integers 5i through 5i+4, where i is its index number. It then returns this value as the process exit status. The parent then waits for them all and outputs the total of the exit statuses, which will be the sum of the integers from 0 to 24 inclusive.
This is an abuse of process exit status, but I think it's fun and a good exercise.

sample solution -- please don't look at the sample solution until you've spent some time on this problem!

2. pwd exercise

Write a program which does most of the work of "pwd", as follows:
First, call lstat() on "." to find out the inode number of the current directory (the "st_ino" member of the struct stat). (Remember that we looked at stat() and lstat() in tutorial five.)

Then we change our current directory to the parent directory, with chdir("..").

Then we want to read the parent directory with opendir()/readdir() (using "opendir(".")" because we've now cd'd there), but unlike in the assignment two directory traversal problem, we are mostly interested in the d_ino value in the struct returned by readdir(). We are looking for the original directory in the parent directory. When we find it, we print its name.

Then call chdir("..") again, and repeat until the inode number is 2, at which point you can print "and then we're at the root directory" (or some such similar message).

So if the real pwd would output "/home/ajr/209/example", your pwd will output

example 
209 
ajr 
home 
and then we're at the root directory.
Except... there are two wrinkles here which affect specifically the linux machines at UTM (and many other computers):

Most of all, there is something strange about the network filesystem mount /student. It is listed under a different inode number in the root directory, as opposed to what lstat() will give you about "." when you are cd'd to /student. There seems to be some unfortunate non-unix semantics with this network file server.

So if you are running this while cd'd to something under your home directory, when you get to /student you will not be able to find its inode number in the parent, and your program will (should) output an appropriate error message to this effect. So at least some of your testing needs to be while cd'd to directories outside your home directory. For example, there are interesting world-readable directory hierarchies under /usr/local/share for you to use (especially /usr/local/share/man) (and there are many others).

Secondly, even in the absence of the funniness around /student, the above algorithm will often stop short of the real root directory, because there are multiple filesystems which are "mounted" in the directory tree, as we discussed when discussing the unix filesystem; and the root directory of any filesystem has inode number 2. We won't fix this for this exercise today; your program will just show you part of the pwd data. All this is to say that you might have to cd into some subdirectories to make the output interesting.

Remember to include all relevant error checking, with appropriate use of perror().

The "real" pwd in linux these days actually uses a totally different algorithm, which is a linux-specific interface for querying the current working directory. But the above is the traditional algorithm. But this means, especially given the nature of /student, that you might want to compare your program's behaviour to mine. So I've put a compiled solution in /student/csc209/tut/08/pwd

(To be clear: You may not use getcwd(). You are to use the above algorithm. An algorithm similar to the above traditionally appears in getcwd(). Your task is to write this algorithm.)

(Note that you don't need to declare the argc and argv parameters to main (and don't need to check usage).)

To submit

First of all, you must run "/student/csc209/present", during the tutorial time, on the console of a tutorial lab workstation.

Then, submit your best version of pwd.c for credit for this tutorial by the end of the day your tutorial is on. To get the mark, it has to perform the basic pwd function using the above algorithm, and it has to compile with "gcc -Wall pwd.c" on the UTM linux machines with no errors or warning messages.

Submit with the command

        /student/csc209/submit tut8 pwd.c
and the other submit commands are as before.