tutorial11: info&answers
CSC 209 tutorial 11 exercises, week of 28 March 2016

Please begin by running "/student/csc209/present". This records your attendance and is required for you to get the marks for the lab.
Exercise one

Here is a demonstration C program, which you can find in /student/csc209/tut/11/bug.c:
#include <stdio.h>
#include <stdlib.h>

int main()
{
    char s[100];

    gets(s);
    if (atoi(s) == 12)
        printf("do something\n");
    printf("in any case, do something else\n");
    return(0);
}
Copy it to your own directory, and compile and run it. (We'll discuss the warning about "gets" in a moment; ignore that at first.)

Try typing the secret value 12, and try typing other numbers.

This seems fine, but it it uses the gets() function. If you did not see the compiler warning about gets(), you omitted the "-Wall" option to gcc. Don't do this. Run it again with "-Wall" to see the warning. You should always compile with "gcc -Wall".

What is wrong with the gets() function? The problem is that it does not take a parameter specifying the size of the array, so it is impossible for it to be written properly and it is impossible for any program which calls it to be correct! If you type more than 99 characters, your program will exceed the 's' array bounds. If you're lucky, your program will simply crash.

1. Cause this program to crash with "segmentation exception".

2. Fix the bug without changing the size of s.

Exercise two

Here is a demonstration C program, which you can find in /student/csc209/tut/11/auth.c:
#include <stdio.h>
#include <string.h>

int main()
{
    extern int authenticate();
    if (authenticate() == 2) {
        printf("Login incorrect.\n");
        return(1);
    }
    printf("Access to secret stuff achieved\n");
    return(0);
}


int authenticate()
{
    char buf[80];
    printf("Password: ");
    if (fgets(buf, sizeof buf, stdin) == NULL)
        return(0);
    else if (strcmp(buf, "sesame\n") == 0)
        return(1);
    else
        return(2);
}
1. Compile and run this. The password is "sesame". Try running it with and without the correct password.

2. Suppose you did not know that the password was sesame but you could see the rest of the source code. Run the program and get it to print "Access to secret stuff achieved" without typing "sesame".

3. Fix the bug.

To submit

First of all, you must run "/student/csc209/present", during the tutorial time, on the console of a tutorial lab workstation.
Submit your fixed bug.c and auth.c programs with commands such as

        /student/csc209/submit tut11 bug.c auth.c
and the other submit commands are as before.
This tutorial will be graded out of 2, one point for each file. To get the point for a file, three things must be true:

you have fixed the bug
your file compiles with "gcc -Wall" with no errors or warning messages
you have not changed anything else which is not related to fixing the bug.